+++ src/Concerns/ValidateableData.php
@@ -3,7 +3,6 @@
 namespace Spatie\LaravelData\Concerns;
 
 use Illuminate\Contracts\Support\Arrayable;
-use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
 use Illuminate\Validation\Validator;
 use Spatie\LaravelData\Resolvers\DataValidationRulesResolver;
@@ -50,9 +49,7 @@ trait ValidateableData
 
     public static function validateAndCreate(Arrayable|array $payload): static
     {
-        if (! $payload instanceof Request) {
-            $payload = static::validate($payload);
-        }
+        static::validate($payload);
 
         return static::from($payload);
     }