+++ tests/Support/EloquentCasts/DataCollectionEloquentCastTest.php
@@ -7,7 +7,9 @@ use function Pest\Laravel\assertDatabaseHas;
 use Spatie\LaravelData\DataCollection;
 
 use Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithCasts;
+
 use Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithCustomCollectionCasts;
+use Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithDefaultCasts;
 use Spatie\LaravelData\Tests\Fakes\SimpleData;
 use Spatie\LaravelData\Tests\Fakes\SimpleDataCollection;
 
@@ -119,4 +121,17 @@ it('retrieves custom data collection', function () {
             new SimpleData('World'),
         ]
     ));
+});
+
+it('loads a custom data collection when nullable argument used and value is null in database', function () {
+    DB::table('dummy_model_with_casts')->insert([
+        'data' => null,
+    ]);
+
+    /** @var \Spatie\LaravelData\Tests\Fakes\Models\DummyModelWithDefaultCasts $model */
+    $model = DummyModelWithDefaultCasts::first();
+
+    expect($model->data_collection)
+        ->toBeInstanceOf(SimpleDataCollection::class)
+        ->toBeEmpty();
 });