+++ CHANGELOG.md
@@ -14,6 +14,11 @@ All notable changes to `laravel-data` will be documented in this file.
 - Allow creating data objects using `from` without parameters
 - Add support for a Dto and Resource object
 
+## 3.5.1 - 2023-05-12
+
+- Add ability to instantiate an eloquent cast when null value using null database values (#425)
+- Revert only use validated data (https://github.com/spatie/laravel-data/pull/438) -> see #432
+
 ## 3.5.0 - 2023-05-05
 
 - Add support for computed values